added better Makefile without printing out confirmations if not true, and also added flag Wconvertions and fixed all of them

+++ libft/libft/ft_strclen.c
@@ -6,13 +6,15 @@
 /*   By: mrantil <mrantil@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/11 20:19:35 by mrantil           #+#    #+#             */
-/*   Updated: 2021/12/11 20:24:00 by mrantil          ###   ########.fr       */
+/*   Updated: 2022/10/31 10:26:56 by mrantil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strclen(char *str, int c)
+#include "libft.h"
+
+size_t	ft_strclen(char *str, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] && str[i] != c)
